refactor: organize data protocols in folders
