+++ src/data/protocols/db/account/index.ts
@@ -1,0 +1,3 @@
+export * from './add-account-repository'
+export * from './load-account-by-email-repository'
+export * from './update-access-token-repository'
