+++ src/data/protocols/db/index.ts
@@ -1,4 +1,2 @@
-export * from './add-account-repository'
-export * from './error-log-repository'
-export * from './load-account-by-email-repository'
-export * from './update-access-token-repository'
+export * from './account'
+export * from './log'
