+++ src/data/protocols/db/log/error-log-repository.ts
