+++ src/data/protocols/db/log/index.ts
@@ -1,0 +1,1 @@
+export * from './error-log-repository'
